packages/server/tests/agents/pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno==1.7.11",
    "anthropic>=0.57.1",
    "duckduckgo-search>=8.1.1",
    "ibm-watsonx-ai>=1.3.36",
    "mcp>=1.12.4",
    "mcp-contextforge-gateway>=0.3.1",
    "ollama>=0.5.3",
    "openai>=1.96.0",
    "python-dotenv>=1.1.1",
]
